Setup for 1.15.2-Fabric

+++ src/main/java/com/irtimaled/bbor/client/renderers/Renderer.java
@@ -96,7 +96,7 @@ public class Renderer {
     }
 
     private void tex(double u, double v) {
-        bufferBuilder.texture(u, v);
+        bufferBuilder.texture((float) u, (float) v);
     }
 
     private void color() {
